Codex: e2e web safari Context https://kittycadworkspace.slack.com/archives/C06L2RHP9SL/p1788484918078329?thread_ts=1788475120.909939&cid=C06L2RHP9SL

+++ e2e/playwright/test-utils.ts
@@ -994,6 +994,9 @@ export async function setup(
       IS_PLAYWRIGHT_KEY,
       TOKEN_PERSIST_KEY,
     }) => {
+      if (window.location.protocol === 'about:') {
+        return
+      }
       localStorage.clear()
       localStorage.setItem(TOKEN_PERSIST_KEY, token)
       localStorage.setItem(settingsKey, settings)
